cratHighlighterSubpages.pl: Don't delete wiki json
Stupid and ugly duplication, but means I can use git to parse changes to the wiki files
This is all largely theoretical, just in case someone else updates something on-wiki so the repo is behind, but it's probably for the best.
Also means I can drop the dependency on the (lovely) File::Compare.