crathighlighter: Line scripts up a bit
Ends up making cratHighlighterSubpages.pl a little more crufty, but the end result is they're more similar so should make it easier for testing, etc.

In theory, I'd *love* to have just one script, but the realities of figuring out when I'm on toolforge or not is more annoying than I'd like.  Although, truth be told, I could just use the `cron` environment freely now, if I commit myself to *not* running it locally via `cron`, which I haven't for ages.  Probably a good idea.