cratHighlighterSubpages.pl: Subroutine date formatting, swap date check order
No need to process ArbCom items if they don't end after us.  The prior start date is necessary in case new Arbs are added early (largely unlikely) but if the term is already done no need for two checks.

As for the formatting, regex overkill, just split and join.  Easy.