crathighlighter: Add basic unit test for ArbCom processing
I switched from the chart template to just reading the members list, but either way, some simple unit test for this is a good idea.  The regex isn't too complicated, nor is the source page, but in theory things might start to creep in.  Best to get this set now, then it should be easier to cover my bases when something needs changing.

Super simple with `Test::More` and `prove`