cratHighlighterSubpages.pl: Move environment stuff into setupEnvironment sub
Not much to get out of the way, so this might not make anything actually more clear, but it gets a few 1s out of the way and, in theory, would be a good stopping point if I felt like `require`ing some modules instead of `use`ing them.  That'd be a speed gain with the -h option, but that's rare and fast enough, so I really prefer the cleanliness of having everything all together, tyvm.