cratHighlighter: Tighten some things
Avoid a few pointless assignations without hurting readability too much.  Plenty of avenues for more such refactorings, but it starts to get pretty sigil soupy.  The worst offender is probably the unnecessary `@pages` array of hashrefs with individual page data from the API, but it was just one use and is *almost* like looking at JSON...