updateModernjs.pl: Fix for UTF8 characters being re-encoded
Basically, MediaWiki will automatically encode into UTF8 -- great, good -- but since I've not marked this as `use utf8;` it was getting re-encoded, which was borking for page titles with non-ascii characters.

At first I wasn't entirely sure why this isn't needed for cratHighlighterSubpages.pl and Alaa, where simply doing `utf8 => '1'` in the API call did the trick, but in hindsight, that's because these are specific page titles being requested and queried, not usernames from a list or content from a page.