proveme.pl: More easily run prove from another directory via cron
`prove` likes being run from the actual directory, and (toolforge's?) cron (can?) suck at getting into a directory, and my tests assume `prove` is being run from the directory, so this is just easier.  stdout is redirect to /dev/null since I just want the error messages.

In theory, there are competing interests here.  I'd like the git actions to be as close as possible to the script running in order to ensure we're running the latest version, and I'd like to run the tests after any git updates, *and* I'd like to have the tests run with enough time for me to make bugfixes before the script is run.  Can't do all that!  There's a case to be made that the test could be right right *after* the script, thereby ensuring I have all the information I need in order to make a potential bugfix.  Food for thought.