AmoryBot::CratHighlighter.pm: More improvements to findLocalGroupMembers
Skip regex entirely in favor of a lookup hash.

Much faster, especially as the regex within `grep` was being called thousands of times (~4700?) and is quite slow, comparatively.  Sad to miss `grep` though, as I do love it.  Still, got a nice `map` in there instead, so it shakes out.