Move botShutoffs to lib
Well, at least the data processing and error production.  The rest is still handled in the main cratHighlighterSubpages.pl script because it has MediaWiki::API and Log::Log4perl; this just processes the data from the query and returns the error messages, but at least it's more appropriately testable.

This is maybe a model for modularizing some of the other stuff that relies on various modules, like MediaWiki::API, Log::Log4perl, JSON::MaybeXS, etc.