11-cmpJSON.t: Remove reliance on processFileData
Holdover from earliest steps of modularization, and redundant to 10-file.t.  A lot of this is there, and was being gone over twice; no need for that.  Maybe this is less effective now, but it's more accurate anyway.

In theory I could rename this to move up in the queue but that doesn't seem necessary?