AmoryBot::CratHighlighter::findLocalGroupMembers: Skip autoconfirmed as well, account for Edit filter
The "Edit filter" user has `sysop` but is not `autoconfirmed`, so we can speed things along by skipping the first *three* user groups for each user, but then need to manually add "Edit filter" to the list of sysops.  That kludge should probably be in the main script, not the library itself?

Maybe one day it'll get fixed: <https://phabricator.wikimedia.org/T212268>