cmpJSON: A bunch of speed improvements
Huge improvement, mostly by checking the length of the arrays first, which should cover most scenarios, and then sorting in place rather than assigning to an array (excessive and only really matters for huge amounts of data and runs, but hey why not).  Also invert the logic of the `$state` check (and do away with that variable itself (which is less clear but fine?)).