cratHighlighter: Remove crontab.crontab and readonly_shim.sh, not being used
No need for crontab since running on kubernetes, and I don't really feel like re-enabling the read-only mode stuff.  It was finalized in 4af42c2, in response to a real issue (<https://phabricator.wikimedia.org/T333477>), but it's rare and I just don't quite care enough.

I think I'll keep the ability to turn off logging for the time being, though (from 886feda), as it's minimal and might be useful down the line.