crathighlighter: Remove tools.amorybot LOGNAME stuff, adjust CRON wording
Now that we're reliably running via toolforge kubernetes, there's no need to keep that in.  It's maybe weird to keep using it via the manually added `LOGNAME`, but not really; need some trigger, so why not that?  Also tweak some of the wording to be less cron specific.

Indeed, keeping both `LOGNAME` and `CRON` allows for better control, and there's no need to eliminate one of them: the former delineates who, and the latter delineates what happens with the output.  I like 'em both.