Add 50-mwUA.t to test the MediaWiki::API object, mainly for UA
Simply copy the subroutine from the main script.  Annoying, but it avoids logging in and `dieNice` (which in turn requires `Log::Log4perl`).  Both of those can be removed, but it also feels silly to import `MediaWiki::API` just for this, and to remove it from the main script?  I dunno.  Anyway, it's tested now.