AmoryBot::CratHighlighter: Use Carp's croak instead of return-ing undef
Better messaging, better handling of actual errors that we're not dealing with in the main script.  Debugging easier?  Tests are pretty robust these days so shouldn't actually be much of an issue, but Better and More Correct.  Also a little tidying up, e.g. some actual (minor) error handling in `findStewardMembers`, dealing with `||` or `or` differences in control flow/assignment.  All exception testing via `Test::Fatal`.